Have a separate RetransmissionBuffer class
