+++ worker/include/RTC/RetransmissionBuffer.hpp
@@ -1,0 +1,65 @@
+#ifndef MS_RTC_RTP_RETRANSMISSION_BUFFER_HPP
+#define MS_RTC_RTP_RETRANSMISSION_BUFFER_HPP
+
+#include "common.hpp"
+#include "RTC/RtpPacket.hpp"
+#include <deque>
+
+namespace RTC
+{
+	// Special container that stores `Item`* elements addressable by their `uint16_t`
+	// sequence number, while only taking as little memory as necessary to store
+	// the range covering a maximum of `MaxRetransmissionDelayForVideoMs` or
+	//  `MaxRetransmissionDelayForAudioMs` ms.
+	class RetransmissionBuffer
+	{
+	public:
+		struct Item
+		{
+			void Reset();
+
+			// Original packet.
+			std::shared_ptr<RTC::RtpPacket> packet{ nullptr };
+			// Correct SSRC since original packet may not have the same.
+			uint32_t ssrc{ 0u };
+			// Correct sequence number since original packet may not have the same.
+			uint16_t sequenceNumber{ 0u };
+			// Correct timestamp since original packet may not have the same.
+			uint32_t timestamp{ 0u };
+			// Last time this packet was resent.
+			uint64_t resentAtMs{ 0u };
+			// Number of times this packet was resent.
+			uint8_t sentTimes{ 0u };
+		};
+
+	public:
+		RetransmissionBuffer(uint16_t maxItems, uint32_t maxRetransmissionDelayMs, uint32_t clockRate);
+		~RetransmissionBuffer();
+
+		Item* Get(uint16_t seq) const;
+		void Insert(RTC::RtpPacket* packet, std::shared_ptr<RTC::RtpPacket>& sharedPacket);
+		void Clear();
+
+	private:
+		void Dump() const;
+		Item* GetOldest() const;
+		Item* GetNewest() const;
+		void RemoveOldest();
+		void RemoveFromFrontAtLeast(uint16_t numItems);
+		void ClearTooOld();
+		bool IsTooOld(uint32_t timestamp, uint32_t newestTimestamp) const;
+		Item* FillItem(
+		  Item* item, RTC::RtpPacket* packet, std::shared_ptr<RTC::RtpPacket>& sharedPacket) const;
+
+	private:
+		// Given as argument.
+		uint16_t maxItems;
+		uint32_t maxRetransmissionDelayMs;
+		uint32_t clockRate;
+		// Others.
+		uint16_t startSeq{ 0u };
+		std::deque<Item*> buffer;
+	};
+} // namespace RTC
+
+#endif
